[refactor] Fixed-width integral types for inside of binary_count_digits

+++ lib/euclidean/euclidean.c
@@ -31,9 +31,9 @@ static uint16_t pattern_offset(uint16_t pattern, uint8_t pattern_len, uint8_t of
 /// Find the length of a binary number by counting bitwise
 static uint8_t binary_count_digits(uint16_t bnry) {
   bool lengthfound = false;
-  int length = 1; // no number can have a length of zero - single 0 has a length of one, but no 1s for the sytem to count
+  uint8_t length = 1; // no number can have a length of zero - single 0 has a length of one, but no 1s for the sytem to count
   for (int q = 32; q >= 0; q--) {
-    int r = (bnry >> q) & 0x01;
+    uint8_t r = (bnry >> q) & 0x01;
     if (r == 1 && lengthfound == false) {
       length = q + 1;
       lengthfound = true;
